chore: migrate user operator schemas to v2 and update imports in related files
