+++ bc_obps/registration/api/v2/_operators/_operator_id/request_access.py
@@ -3,7 +3,7 @@ from uuid import UUID
 from common.permissions import authorize
 from django.http import HttpRequest
 from registration.constants import OPERATOR_TAGS
-from registration.schema.v1.user_operator import RequestAccessOut
+from registration.schema.v2.user_operator import RequestAccessOut
 from common.api.utils import get_current_user_guid
 from registration.schema import Message
 from registration.api.router import router
