+++ bc_obps/registration/api/v2/_user_operators/_current/access_requests.py
@@ -5,7 +5,7 @@ from registration.constants import USER_OPERATOR_TAGS
 from registration.models import UserOperator
 from service.data_access_service.user_operator_service import UserOperatorDataAccessService
 from django.db.models import QuerySet
-from registration.schema.v1 import ExternalDashboardUsersTileData
+from registration.schema.v2.user_operator import ExternalDashboardUsersTileData
 from registration.schema.generic import Message
 from typing import List, Literal, Tuple
 from registration.api.router import router
