+++ bc_obps/registration/schema/v2/parent_operator.py
@@ -33,6 +33,34 @@ class ParentOperatorIn(ModelSchema):
         fields = ["id"]
         populate_by_name = True
 
+class ParentOperatorOutV1(ModelSchema):
+    """
+    Schema for the Parent Operator part of the user operator form
+    """
+
+    po_legal_name: str = Field(..., alias="legal_name")
+    po_trade_name: Optional[str] = Field("", alias="trade_name")
+    po_cra_business_number: int = Field(..., alias="cra_business_number")
+    po_bc_corporate_registry_number: str = Field(..., alias="bc_corporate_registry_number")
+    po_business_structure: str = Field(..., alias="business_structure.name")
+    po_website: Optional[str] = Field("", alias="website")
+    po_physical_street_address: str = Field(..., alias="physical_address.street_address")
+    po_physical_municipality: str = Field(..., alias="physical_address.municipality")
+    po_physical_province: str = Field(..., alias="physical_address.province")
+    po_physical_postal_code: str = Field(..., alias="physical_address.postal_code")
+    po_mailing_address_same_as_physical: bool
+    po_mailing_street_address: Optional[str] = Field(None, alias="mailing_address.street_address")
+    po_mailing_municipality: Optional[str] = Field(None, alias="mailing_address.municipality")
+    po_mailing_province: Optional[str] = Field(None, alias="mailing_address.province")
+    po_mailing_postal_code: Optional[str] = Field(None, alias="mailing_address.postal_code")
+
+    @staticmethod
+    def resolve_po_mailing_address_same_as_physical(obj: ParentOperator) -> bool:
+        return obj.mailing_address_id == obj.physical_address_id
+
+    class Meta:
+        model = ParentOperator
+        fields = ["operator_index"]
 
 class ParentOperatorOut(ModelSchema):
     """
