+++ bc_obps/service/tests/test_user_operator_service_v2.py
@@ -5,9 +5,8 @@ from model_bakery import baker
 
 from registration.constants import UNAUTHORIZED_MESSAGE
 from registration.models import Operator, User, UserOperator, Contact
-from registration.schema.v1.user_operator import UserOperatorStatusUpdate
 from registration.schema.v2.operator import OperatorIn
-from registration.schema.v2.user_operator import UserOperatorFilterSchema
+from registration.schema.v2.user_operator import UserOperatorFilterSchema, UserOperatorStatusUpdate
 from service.user_operator_service_v2 import UserOperatorServiceV2
 
 pytestmark = pytest.mark.django_db
